modules: extract base path for conditional
